refactor : moving logging and database to the project's infra layer
